setup: `Config.java`: optimization (partly)
also:
- Move `Config.java` to `setup` package
