+++ app/src/main/java/com/osfans/trime/ime/core/Trime.java
@@ -55,7 +55,7 @@ import android.widget.PopupWindow;
 
 import com.osfans.trime.Candidate;
 import com.osfans.trime.Composition;
-import com.osfans.trime.Config;
+import com.osfans.trime.setup.Config;
 import com.osfans.trime.Effect;
 import com.osfans.trime.Event;
 import com.osfans.trime.Function;
