+++ app/src/main/java/com/osfans/trime/settings/components/ThemePickerDialog.kt
@@ -5,7 +5,7 @@ import android.app.ProgressDialog
 import android.content.Context
 import android.os.IBinder
 import android.view.WindowManager
-import com.osfans.trime.Config
+import com.osfans.trime.setup.Config
 import com.osfans.trime.R
 import com.osfans.trime.ime.core.Trime
 import kotlinx.coroutines.*
